backend/requirements.txt
# Django e extensões
Django==5.0.8
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5

# Database
psycopg[binary,pool]==3.1.19

# Cache e Session
django-redis==5.4.0
redis==5.0.7

# Celery
celery==5.4.0
django-celery-beat==2.6.0
django-celery-results==2.5.1

# Configuration Management
dynaconf==3.2.5
python-decouple==3.8

# Security
django-ratelimit==4.1.0
django-axes==6.1.1

# File handling
Pillow==10.4.0
python-docx==1.1.0
PyPDF2==3.0.1
python-magic==0.4.27

# API Documentation
drf-spectacular==0.27.2

# Monitoring and Logging
sentry-sdk[django]==2.8.0

# Development tools
django-debug-toolbar==4.4.6
factory-boy==3.3.0
faker==25.8.0

# Testing
pytest==8.2.2
pytest-django==4.8.0
pytest-cov==5.0.0
pytest-mock==3.14.0

# Code quality
black==24.4.2
isort==5.13.2
flake8==7.1.0
mypy==1.10.1
django-stubs==5.0.2

# Production server
gunicorn==22.0.0
whitenoise==6.7.0

# Utilities
requests==2.32.3
python-dateutil==2.9.0.post0
pytz==2024.1
openpyxl==3.1.5

# Internationalization
django-modeltranslation==0.18.11

# File storage (optional - for S3)
django-storages[s3]==1.14.4
boto3==1.34.144

# Environment variables
python-dotenv==1.0.1

# Validation
cerberus==1.3.5
marshmallow==3.21.3

# Background tasks
django-rq==2.10.2

# API clients
httpx==0.27.0

# Data processing
pandas==2.2.2
numpy==2.0.0

# OCR (if needed)
pytesseract==0.3.10

# Cryptography
cryptography==42.0.8

